+++ packages/auth/__tests__/providers/cognito/fetchMFAPreference.test.ts
@@ -4,10 +4,11 @@
 import { Amplify, fetchAuthSession } from '@aws-amplify/core';
 import { decodeJWT } from '@aws-amplify/core/internals/utils';
 
-import { getUser } from '../../../src/providers/cognito/utils/clients/CognitoIdentityProvider';
 import { AuthError } from '../../../src/errors/AuthError';
 import { fetchMFAPreference } from '../../../src/providers/cognito/apis/fetchMFAPreference';
 import { GetUserException } from '../../../src/providers/cognito/types/errors';
+import { createGetUserClient } from '../../../src/foundation/factories/serviceClients/cognitoIdentityProvider';
+import { createCognitoUserPoolEndpointResolver } from '../../../src/providers/cognito/factories';
 
 import { getMockError, mockAccessToken } from './testUtils/data';
 import { setUpGetConfig } from './testUtils/setUpGetConfig';
@@ -17,19 +18,25 @@ jest.mock('@aws-amplify/core', () => ({
 	Amplify: { getConfig: jest.fn(() => ({})) },
 }));
 jest.mock(
-	'../../../src/providers/cognito/utils/clients/CognitoIdentityProvider',
+	'../../../src/foundation/factories/serviceClients/cognitoIdentityProvider',
 );
+jest.mock('../../../src/providers/cognito/factories');
 
 describe('fetchMFAPreference', () => {
 	// assert mocks
 	const mockFetchAuthSession = jest.mocked(fetchAuthSession);
-	const mockGetUser = jest.mocked(getUser);
+	const mockGetUser = jest.fn();
+	const mockCreateGetUserClient = jest.mocked(createGetUserClient);
+	const mockCreateCognitoUserPoolEndpointResolver = jest.mocked(
+		createCognitoUserPoolEndpointResolver,
+	);
 
 	beforeAll(() => {
 		setUpGetConfig(Amplify);
 		mockFetchAuthSession.mockResolvedValue({
 			tokens: { accessToken: decodeJWT(mockAccessToken) },
 		});
+		mockCreateGetUserClient.mockReturnValue(mockGetUser);
 	});
 
 	afterEach(() => {
@@ -100,6 +107,34 @@ describe('fetchMFAPreference', () => {
 		});
 		const resp = await fetchMFAPreference();
 		expect(resp).toEqual({});
+	});
+
+	it('invokes mockCreateCognitoUserPoolEndpointResolver with expected endpointOverride', async () => {
+		const expectedUserPoolEndpoint = 'https://my-custom-endpoint.com';
+		jest.mocked(Amplify.getConfig).mockReturnValueOnce({
+			Auth: {
+				Cognito: {
+					userPoolClientId: '111111-aaaaa-42d8-891d-ee81a1549398',
+					userPoolId: 'us-west-2_zzzzz',
+					identityPoolId: 'us-west-2:xxxxxx',
+					userPoolEndpoint: expectedUserPoolEndpoint,
+				},
+			},
+		});
+
+		mockGetUser.mockResolvedValueOnce({
+			UserAttributes: [],
+			Username: 'XXXXXXXX',
+			PreferredMfaSetting: 'SMS_MFA',
+			UserMFASettingList: ['SMS_MFA', 'SOFTWARE_TOKEN_MFA', 'EMAIL_OTP'],
+			$metadata: {},
+		});
+
+		await fetchMFAPreference();
+
+		expect(mockCreateCognitoUserPoolEndpointResolver).toHaveBeenCalledWith({
+			endpointOverride: expectedUserPoolEndpoint,
+		});
 	});
 
 	it('should throw an error when service returns an error response', async () => {