+++ packages/auth/__tests__/providers/cognito/resendSignUpCode.test.ts
@@ -7,7 +7,8 @@ import { resendSignUpCode } from '../../../src/providers/cognito';
 import { AuthValidationErrorCode } from '../../../src/errors/types/validation';
 import { AuthError } from '../../../src/errors/AuthError';
 import { ResendConfirmationException } from '../../../src/providers/cognito/types/errors';
-import { resendConfirmationCode } from '../../../src/providers/cognito/utils/clients/CognitoIdentityProvider';
+import { createResendConfirmationCodeClient } from '../../../src/foundation/factories/serviceClients/cognitoIdentityProvider';
+import { createCognitoUserPoolEndpointResolver } from '../../../src/providers/cognito/factories';
 
 import { authAPITestParams } from './testUtils/authApiTestParams';
 import { getMockError } from './testUtils/data';
@@ -22,13 +23,20 @@ jest.mock('@aws-amplify/core/internals/utils', () => ({
 	isBrowser: jest.fn(() => false),
 }));
 jest.mock(
-	'../../../src/providers/cognito/utils/clients/CognitoIdentityProvider',
+	'../../../src/foundation/factories/serviceClients/cognitoIdentityProvider',
 );
+jest.mock('../../../src/providers/cognito/factories');
 
 describe('resendSignUpCode', () => {
 	const { user1 } = authAPITestParams;
 	// assert mocks
-	const mockResendConfirmationCode = resendConfirmationCode as jest.Mock;
+	const mockResendConfirmationCode = jest.fn();
+	const mockCreateResendConfirmationCodeClient = jest.mocked(
+		createResendConfirmationCodeClient,
+	);
+	const mockCreateCognitoUserPoolEndpointResolver = jest.mocked(
+		createCognitoUserPoolEndpointResolver,
+	);
 
 	beforeAll(() => {
 		setUpGetConfig(Amplify);
@@ -37,6 +45,9 @@ describe('resendSignUpCode', () => {
 	beforeEach(() => {
 		mockResendConfirmationCode.mockResolvedValue(
 			authAPITestParams.resendSignUpClientResult,
+		);
+		mockCreateResendConfirmationCodeClient.mockReturnValueOnce(
+			mockResendConfirmationCode,
 		);
 	});
 
@@ -61,6 +72,26 @@ describe('resendSignUpCode', () => {
 			},
 		);
 		expect(mockResendConfirmationCode).toHaveBeenCalledTimes(1);
+	});
+
+	it('invokes createCognitoUserPoolEndpointResolver with expected endpointOverride', async () => {
+		const expectedUserPoolEndpoint = 'https://my-custom-endpoint.com';
+		jest.mocked(Amplify.getConfig).mockReturnValueOnce({
+			Auth: {
+				Cognito: {
+					userPoolClientId: '111111-aaaaa-42d8-891d-ee81a1549398',
+					userPoolId: 'us-west-2_zzzzz',
+					identityPoolId: 'us-west-2:xxxxxx',
+					userPoolEndpoint: expectedUserPoolEndpoint,
+				},
+			},
+		});
+		await resendSignUpCode({
+			username: user1.username,
+		});
+		expect(mockCreateCognitoUserPoolEndpointResolver).toHaveBeenCalledWith({
+			endpointOverride: expectedUserPoolEndpoint,
+		});
 	});
 
 	it('should throw an error when username is empty', async () => {