+++ packages/auth/__tests__/providers/cognito/updateMFAPreference.test.ts
@@ -8,11 +8,12 @@ import {
 	UpdateMFAPreferenceInput,
 	updateMFAPreference,
 } from '../../../src/providers/cognito';
-import { setUserMFAPreference } from '../../../src/providers/cognito/utils/clients/CognitoIdentityProvider';
 import { AuthError } from '../../../src/errors/AuthError';
 import { SetUserMFAPreferenceException } from '../../../src/providers/cognito/types/errors';
 import { getMFASettings } from '../../../src/providers/cognito/apis/updateMFAPreference';
 import { MFAPreference } from '../../../src/providers/cognito/types';
+import { createSetUserMFAPreferenceClient } from '../../../src/foundation/factories/serviceClients/cognitoIdentityProvider';
+import { createCognitoUserPoolEndpointResolver } from '../../../src/providers/cognito/factories';
 
 import { getMockError, mockAccessToken } from './testUtils/data';
 import { setUpGetConfig } from './testUtils/setUpGetConfig';
@@ -28,8 +29,9 @@ jest.mock('@aws-amplify/core/internals/utils', () => ({
 	isBrowser: jest.fn(() => false),
 }));
 jest.mock(
-	'../../../src/providers/cognito/utils/clients/CognitoIdentityProvider',
+	'../../../src/foundation/factories/serviceClients/cognitoIdentityProvider',
 );
+jest.mock('../../../src/providers/cognito/factories');
 
 // generates all preference permutations
 const generateUpdateMFAPreferenceOptions = () => {
@@ -63,10 +65,18 @@ const generateUpdateMFAPreferenceOptions = () => {
 	return generatePermutations(mfaKeys, mfaPreferenceTypes);
 };
 
+const mfaChoices = generateUpdateMFAPreferenceOptions();
+
 describe('updateMFAPreference', () => {
 	// assert mocks
 	const mockFetchAuthSession = fetchAuthSession as jest.Mock;
-	const mockSetUserMFAPreference = setUserMFAPreference as jest.Mock;
+	const mockSetUserMFAPreference = jest.fn();
+	const mockCreateSetUserMFAPreferenceClient = jest.mocked(
+		createSetUserMFAPreferenceClient,
+	);
+	const mockCreateCognitoUserPoolEndpointResolver = jest.mocked(
+		createCognitoUserPoolEndpointResolver,
+	);
 
 	beforeAll(() => {
 		setUpGetConfig(Amplify);
@@ -77,14 +87,18 @@ describe('updateMFAPreference', () => {
 
 	beforeEach(() => {
 		mockSetUserMFAPreference.mockResolvedValue({});
+		mockCreateSetUserMFAPreferenceClient.mockReturnValueOnce(
+			mockSetUserMFAPreference,
+		);
 	});
 
 	afterEach(() => {
 		mockSetUserMFAPreference.mockReset();
 		mockFetchAuthSession.mockClear();
+		mockCreateSetUserMFAPreferenceClient.mockClear();
 	});
 
-	it.each(generateUpdateMFAPreferenceOptions())(
+	it.each(mfaChoices)(
 		'should update with email $email, sms $sms, and totp $totp',
 		async mfaChoice => {
 			const { totp, sms, email } = mfaChoice;
@@ -103,6 +117,25 @@ describe('updateMFAPreference', () => {
 			);
 		},
 	);
+
+	it('invokes mockCreateCognitoUserPoolEndpointResolver with expected endpointOverride', async () => {
+		const expectedUserPoolEndpoint = 'https://my-custom-endpoint.com';
+		jest.mocked(Amplify.getConfig).mockReturnValueOnce({
+			Auth: {
+				Cognito: {
+					userPoolClientId: '111111-aaaaa-42d8-891d-ee81a1549398',
+					userPoolId: 'us-west-2_zzzzz',
+					identityPoolId: 'us-west-2:xxxxxx',
+					userPoolEndpoint: expectedUserPoolEndpoint,
+				},
+			},
+		});
+		await updateMFAPreference(mfaChoices[0]);
+
+		expect(mockCreateCognitoUserPoolEndpointResolver).toHaveBeenCalledWith({
+			endpointOverride: expectedUserPoolEndpoint,
+		});
+	});
 
 	it('should throw an error when service returns an error response', async () => {
 		expect.assertions(2);