+++ packages/auth/src/providers/cognito/apis/confirmSignIn.ts
@@ -17,7 +17,7 @@ import {
 } from '../utils/signInStore';
 import { AuthError } from '../../../errors/AuthError';
 import {
-	getNewDeviceMetatada,
+	getNewDeviceMetadata,
 	getSignInResult,
 	getSignInResultFromError,
 	handleChallengeName,
@@ -27,12 +27,12 @@ import { assertValidationError } from '../../../errors/utils/assertValidationErr
 import { AuthValidationErrorCode } from '../../../errors/types/validation';
 import { AuthErrorCodes } from '../../../common/AuthErrorStrings';
 import { cacheCognitoTokens } from '../tokenProvider/cacheTokens';
+import { tokenOrchestrator } from '../tokenProvider';
+import { dispatchSignedInHubEvent } from '../utils/dispatchSignedInHubEvent';
 import {
 	ChallengeName,
 	ChallengeParameters,
-} from '../utils/clients/CognitoIdentityProvider/types';
-import { tokenOrchestrator } from '../tokenProvider';
-import { dispatchSignedInHubEvent } from '../utils/dispatchSignedInHubEvent';
+} from '../../../foundation/factories/serviceClients/cognitoIdentityProvider/types';
 
 /**
  * Continues or completes the sign in process when required by the initial call to `signIn`.
@@ -113,11 +113,12 @@ export async function confirmSignIn(
 			await cacheCognitoTokens({
 				username,
 				...AuthenticationResult,
-				NewDeviceMetadata: await getNewDeviceMetatada(
-					authConfig.userPoolId,
-					AuthenticationResult.NewDeviceMetadata,
-					AuthenticationResult.AccessToken,
-				),
+				NewDeviceMetadata: await getNewDeviceMetadata({
+					userPoolId: authConfig.userPoolId,
+					userPoolEndpoint: authConfig.userPoolEndpoint,
+					newDeviceMetadata: AuthenticationResult.NewDeviceMetadata,
+					accessToken: AuthenticationResult.AccessToken,
+				}),
 				signInDetails,
 			});